+++ requirement.txt
@@ -1,11 +1,8 @@
-
 opencv-contrib-python==4.4.0.44
 flask==1.1.2
 imutils==0.5.3
 scikit-learn==0.23.2
-
 tensorflow==2.3.1
-filetype==1.0.7
 pillow==7.2.0
 Keras==2.4.3
-mtcnn==0.1.0
+moviepy==1.0.3
