+++ webapp/cfg/classes.names
@@ -1,2 +1,2 @@
-mask
-no_mask
+crying_baby
+others
